requirements.txt
selenium==4.9.1
webdriver_manager==4.0.0
paho-mqtt==2.1.0